DoorMovesWoah
No buttons or triggers needed to activate the door yet, but it opens for the player if they press X on the controller. The only problem now is that the door animation is set to the position, so I might need to change the position of the animation for each new door?